Create info
In this game, the player controls a car and tries to reach the finish line. The game board is represented by a series of dashes ("-"), with the car represented by the letter "C". The player can enter "l" to move the car left or "r" to move the car right. The game continues until the car reaches the finish line.

Please note that this is a basic implementation, and you can add more features, such as obstacles, scoring, levels, and graphics, to make the game more interesting.